# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Poop Recorder API
  version: 0.1.0
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      responses:
        '200': { description: OK }
  /auth/register:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '201': { description: Created }
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: OK }
  /entries:
    get:
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                timestampMinute: { type: string, format: date-time }
                bristolType: { type: integer }
                smellScore: { type: integer }
                color: { type: string }
                volume: { type: string }
                symptoms: { type: array, items: { type: string } }
                notes: { type: string }
      responses:
        '201': { description: Created }
  /analytics/frequency:
    get:
      parameters:
        - in: query
          name: period
          schema: { type: string, enum: [week, month, quarter] }
      responses:
        '200': { description: OK }
  /analytics/score:
    get:
      responses:
        '200': { description: OK }
  /alerts:
    get:
      responses:
        '200': { description: OK }
  /privacy/export:
    post:
      responses:
        '200': { description: OK }
  /privacy/delete:
    post:
      responses:
        '200': { description: OK }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT